[SLP][modularisation][NFC] Add SLPTree.h; move ReductionVectorPart
Create the internal header SLPVectorizer/SLPTree.h and move the
BoUpSLP-independent ReductionVectorPart struct and the MinScheduleRegionSize
constant into it. Prepares for relocating the BoUpSLP class declaration into
this header in a follow-up. NFC.

Part of the SLPVectorizer.cpp modularization effort:
https://discourse.llvm.org/t/modularizing-slpvectorizer-cpp/90922
